rename runDev -> run-dev
