+++ src/lib/run-dev.js
@@ -21,7 +21,7 @@ const runDevLocal = require('./run-dev-local')
 const BuildActions = require('@adobe/aio-lib-runtime').buildActions
 const DeployActions = require('@adobe/aio-lib-runtime').deployActions
 const utils = require('./app-helper')
-const EventPoller = require('../lib/poller')
+const EventPoller = require('./poller')
 const execa = require('execa')
 const chokidar = require('chokidar')
 let running = false
